feat: re-order the versions so that the newest version is tested first
Travis runs jobs from top to bottom, and the newest nodes are both the most important ones and the fastest ones, so to get the fastest feedback, you want to see those results first (and you want them to thus finish first). This is more important when you have more than 5 builds (since 5 is the current throttling threshold for travis) but can also happen when you have other projects in the same org running.

Signed-off-by: Dominykas Blyžė <hello@dominykas.com>